requirements.txt
python-telegram-bot==22.6
psycopg2-binary